m8-common: Pin hex'ed sensors HAL
HTC One M8 exposes a custom sensor "HTC Gesture sensor" with type 27,
which is used for SENSOR_TYPE_DEVICE_ORIENTATION sensors. This is now
an issue in Pie, since it is a one-shot sensor and cannot be
registered by the WindowOrientationListener.

Workaround the issue by replacing all the occurrences of 0x1B
(the stock value of 'type', which is a member of 'sensor_t' struct)
to 0x10001 in the sensors HAL.

Alternative to the original solution:

  https://review.lineageos.org/c/LineageOS/android_frameworks_base/+/242737

Change-Id: I99ac93ec485d7c349eb705a4e953e5410d1ae025